# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 골드레슬링 API
  description: 골드레슬링 체육관 회원 관리·예약 시스템 백엔드 API
  version: v0.0.1
servers:
- url: /
...
paths:
...
  /api/admin/batch/inactivity-runs:
    get:
      tags:
      - admin-batch
      summary: 최근 배치 실행 이력 목록 (시작 시각 내림차순)
      description: '"어젯밤 배치가 돌았는가"를 실행 이력으로 확인하는 창구다(D-108). 결과가 비어 있거나 가장 최근 실행이
        어제가 아니면 cron이 멈춘 것이다.'
      operationId: listInactivityRuns
      parameters:
      - name: limit
        in: query
        description: 조회 건수. 1..100을 벗어나면 그 범위로 보정된다
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: 조회 성공
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatchExecutionResponse"
        "401":
          description: 인증되지 않음 (UNAUTHENTICATED)
        "403":
          description: 관리자 권한이 아님 (ACCESS_DENIED)
    post:
      tags:
      - admin-batch
      summary: "2주 미사용 자동 차감 배치 실행을 접수한다 (비동기, 202)"
      description: "이 호출은 배치를 **시작만** 하고 즉시 202를 반환한다 — 응답 본문의 status는 RUNNING이고 finishedAt은\
        \ null이다. 진행 상태와 최종 결과는 Location 헤더가 가리키는 단건 조회(GET /api/admin/batch/inactivity-runs/{batchExecutionId})로\
        \ 폴링한다. **실행 중에 다시 호출하면 409(BATCH_ALREADY_RUNNING)로 거부되므로 재시도가 안전하다** — 동시\
        에 실행 중일 수 있는 배치는 DB 제약으로 항상 최대 1건이다(D-117). **순차 재실행도 안전하다** — 상태 기반 부족분 계\
        산(D-106)이 이미 존재하는 차감을 다시 만들지 않으므로 앞 실행이 끝난 뒤 다시 호출하면 그 실행의 deductedCount는\
        \ 0이 된다. 배치가 며칠 중단됐다면 밀린 주기를 이어받되 **1회 실행당 회원 1명에게서 최대 1회**만 차감한다(D-119) —\
        \ 나머지는 다음 실행들이 이어받는다."
      operationId: runInactivityBatch
      responses:
        "202":
          description: "실행 접수됨 (status = RUNNING, finishedAt = null)"
          headers:
            Location:
              description: 이 실행의 진행 상태 조회 경로
              style: simple
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchExecutionResponse"
        "401":
          description: 인증되지 않음 (UNAUTHENTICATED)
        "403":
          description: 관리자 권한이 아님 (ACCESS_DENIED)
        "409":
          description: 이미 실행 중인 배치가 있어 거부됨 (BATCH_ALREADY_RUNNING)
...
  /api/admin/batch/inactivity-runs/{batchExecutionId}:
    get:
      tags:
      - admin-batch
      summary: 배치 실행 1건의 진행 상태·결과 조회
      description: 202 응답의 Location이 가리키는 폴링 대상이다. status가 RUNNING이면 아직 실행 중이며 집계
        수치는 확정 전 값(0)이고 finishedAt은 null이다. 종료되면 SUCCESS · PARTIAL_FAILURE(일부 회원 실패)
        · FAILED(실행 전체 실패) 중 하나가 된다.
      operationId: getInactivityRun
      parameters:
      - name: batchExecutionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchExecutionResponse"
        "401":
          description: 인증되지 않음 (UNAUTHENTICATED)
        "403":
          description: 관리자 권한이 아님 (ACCESS_DENIED)
        "404":
          description: 그 id의 실행 이력이 없음 (BATCH_EXECUTION_NOT_FOUND)
components:
  schemas:
...
    BatchExecutionResponse:
      type: object
      description: 배치 실행 결과
      properties:
        batchExecutionId:
          type: integer
          format: int64
          description: 배치 실행 이력 ID
        trigger:
          type: string
          description: 트리거 종류
          enum:
          - SCHEDULED
          - MANUAL
        triggeredByAdminId:
          type:
          - integer
          - "null"
          format: int64
          description: 수동 실행한 관리자 ID — SCHEDULED면 null
        startedAt:
          type: string
          format: date-time
          description: 실행 시작 시각
        finishedAt:
          type:
          - string
          - "null"
          format: date-time
          description: 실행 종료 시각 — 아직 실행 중(RUNNING)이면 null
        processedMemberCount:
          type: integer
          format: int32
          description: 차감 대상 후보로 조회된 회원 수
        deductedCount:
          type: integer
          format: int32
          description: 실제 차감된 건수
        skippedCount:
          type: integer
          format: int32
          description: 대상 소진 등으로 스킵된 건수 — 다음 실행이 상태 기반으로 이어받는다
        status:
          type: string
          description: 실행 결과 — RUNNING이면 아직 실행 중이며 집계는 확정 전 값(0)이다
          enum:
          - RUNNING
          - SUCCESS
          - PARTIAL_FAILURE
          - FAILED
        errorSummary:
          type:
          - string
          - "null"
          description: "회원 단위 실패 요약(예외 종류만, 메시지는 담지 않는다) — 실패가 없으면 null"
...